@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.18.6 <3.0.0'
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -14,17 +14,17 @@ dependencies:
 
   cupertino_icons: ^1.0.2
   flutter_easyloading: ^3.0.5
-  go_router: ^6.5.7
-  google_fonts: ^4.0.3
+  go_router: ^14.8.0
+  google_fonts: ^6.2.1
   provider: ^6.0.5
-  window_manager: ^0.2.7
+  window_manager: ^0.4.3
   font_awesome_flutter: ^10.4.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.0
+  flutter_lints: ^5.0.0
 
 flutter:
 
